Refactor predicates and fix child::node()
Transform predicate to interfaces that can implement the String() method
for easier debugging. Also, this allows declarative definition of
predicates and better composition

This allows to fix the issue with child::node() where the node type was
checked in two places. It end up excluding text nodes. Now, with predicate
types it is possible to check node type only once and keep text nodes in
the selection.

Fixes #13