Add `transformFunctionQuery` and its first usage xpath 2.0's `reverse()`
XPath 2.0 introduces a number of functions that returns a node-set for a given input
node-set. That's different from XPath 1.0 functions. Introduce `transformFunctionQuery`
which performs a mapping/transform on a given node-set input (NodeNavigator) and returns
a new node-set. And coming with it is the first application `reverse()`. Plus unit test
coverage.

For more details, check out the discussion in https://github.com/antchfx/xpath/issues/45.